requirements.txt
# Microsoft Agent Framework SDK - New unified framework
agent-framework

# Core dependencies
aiohttp
databricks-sdk
python-dotenv
azure-identity
azure-ai-projects
azure.storage.blob

# Microsoft Agents packages (compatible versions)
microsoft-agents-hosting-core==0.4.0
microsoft-agents-activity==0.4.0
microsoft-agents-hosting-aiohttp==0.4.0
microsoft-agents-authentication-msal==0.4.0
microsoft-agents-hosting-teams==0.4.0

# Azure AI Agents (for model compatibility)
azure-ai-agents==1.2.0b5